+++ Disk/Repository/Interface/IPatientRepository.cs
@@ -5,7 +5,7 @@ namespace Disk.Repository.Interface
 {
     public interface IPatientRepository : ICrudRepository<Patient> 
     { 
-        int GetPatientsCount();
+        long GetPatientsCount();
         ICollection<Patient> GetPatientsByFullname(string name, string surname, string patronymic);
         ICollection<Patient> GetPatientsPage(int pageNum, int patientsPerPage);
     }
